requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
psycopg2-binary
numpy
pillow
docker
pandas
PyYAML
